+++ src/extensions/log-extension.h
@@ -109,6 +109,21 @@ public:
         std::string formattedMessage;
     };
 
+    ///
+    /// \brief Represents typed level. Statically cast Data::level to this
+    /// for readibility purposes
+    ///
+    enum class Level : unsigned int
+    {
+        Trace = 2,
+        Debug = 4,
+        Fatal = 8,
+        Error = 16,
+        Warning = 32,
+        Verbose = 64,
+        Info = 128
+    };
+
     explicit LogExtension(const std::string& id);
     virtual ~LogExtension() = default;