+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 # Change Log
 
 ## [Unreleased]
-### Updates
+### Extensions API
+- Added `LogExtension::Level` enum for readability of `LogExtension::Data::level`
+
+### Extensions Updates
 - Log message when execution of extension finished
 
 ## [2.1.0] - 07-03-2018